+++ docs/implementors/core/marker/trait.Copy.js
@@ -1,5 +1,5 @@
 (function() {var implementors = {};
-implementors["my_iot"] = [{text:"impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/marker/trait.Copy.html\" title=\"trait core::marker::Copy\">Copy</a> for <a class=\"enum\" href=\"my_iot/value/enum.PointOfTheCompass.html\" title=\"enum my_iot::value::PointOfTheCompass\">PointOfTheCompass</a>",synthetic:false,types:["my_iot::value::PointOfTheCompass"]},];
+implementors["my_iot"] = [{text:"impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/marker/trait.Copy.html\" title=\"trait core::marker::Copy\">Copy</a> for <a class=\"enum\" href=\"my_iot/reading/enum.Type.html\" title=\"enum my_iot::reading::Type\">Type</a>",synthetic:false,types:["my_iot::reading::Type"]},{text:"impl <a class=\"trait\" href=\"https://doc.rust-lang.org/nightly/core/marker/trait.Copy.html\" title=\"trait core::marker::Copy\">Copy</a> for <a class=\"enum\" href=\"my_iot/value/enum.PointOfTheCompass.html\" title=\"enum my_iot::value::PointOfTheCompass\">PointOfTheCompass</a>",synthetic:false,types:["my_iot::value::PointOfTheCompass"]},];
 
             if (window.register_implementors) {
                 window.register_implementors(implementors);