+++ docs/my_iot/reading/sidebar-items.js
@@ -1,1 +1,1 @@
-initSidebarItems({"struct":[["Reading","A sensor reading."]]});
+initSidebarItems({"constant":[["_IMPL_DESERIALIZE_FOR_Type",""]],"enum":[["Type","Message type."]],"struct":[["Message","Services use messages to exchange sensor readings between each other."],["Reading","Single sensor reading."]]});